fix(#103): 삭제 기능 및 권한 확인 기능 분리

+++ src/main/java/hello/cluebackend/domain/classroom/service/ClassRoomService.java
@@ -1,5 +1,6 @@
 package hello.cluebackend.domain.classroom.service;
 
+import hello.cluebackend.domain.assignment.exception.AccessDeniedException;
 import hello.cluebackend.domain.classroom.domain.ClassRoom;
 import hello.cluebackend.domain.classroom.domain.repository.ClassRoomRepository;
 import hello.cluebackend.domain.classroom.controller.dto.ClassRoomAllInfoDto;
@@ -21,7 +22,6 @@ import org.springframework.stereotype.Service;
 
 import java.util.Comparator;
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
@@ -32,6 +32,7 @@ public class ClassRoomService {
   private final ClassRoomRepository classRoomRepository;
   private final UserRepository userRepository;
 
+  // 내가 속한 모든 교실 조회 (전체)
   public List<ClassRoomCardDto> findMyClassRoomById(UUID userId) {
     List<ClassRoomUser> classRoomUsers = classRoomUserRepository.findByUser_UserId(userId);
     return classRoomUsers.stream()
@@ -41,6 +42,7 @@ public class ClassRoomService {
             .toList();
   }
 
+  // 교실 생성 (선생)
   @Transactional
   public void createClassRoom(ClassRoomDto classRoomDTO, UUID userId) {
     classRoomDTO.generateCode();
@@ -56,6 +58,7 @@ public class ClassRoomService {
     classRoomUserRepository.save(classRoomUser);
   }
 
+  // 교실 참여 (전체)
   public void joinClassRoom(UUID userId, String code) {
     ClassRoom findClassRoom = classRoomRepository.findByCode(code).orElseThrow(() -> new IllegalArgumentException("classroom not found"));
     UserEntity findUser = userRepository.findById(userId).orElseThrow(() -> new IllegalArgumentException("user not found"));
@@ -66,17 +69,21 @@ public class ClassRoomService {
     classRoomUserRepository.save(classRoomUser);
   }
 
-  public ClassRoomDto findById(UUID userId, UUID classRoomId) {
-  ClassRoom findClassRoom = classRoomRepository.findByIdWithTeachers(classRoomId);
-  return findClassRoom.toDTO();
+  // 교실 단일 조회 (전체)
+  public ClassRoom findById(UUID userId, UUID classId) {
+    validateInClassRoom(userId, classId);
+    return classRoomRepository.findByIdWithTeachers(classId);
   }
 
-  public void updateClassRoom(UUID classId, ClassRoomDto classRoomDTO) {
+  // 교실 수정 (선생)
+  public void updateClassRoom(UUID classId, UUID userId, ClassRoomDto classRoomDTO) {
+    validateOwner(userId, classId);
     ClassRoom findClassRoom =  classRoomRepository.findById(classId).orElseThrow(() -> new IllegalArgumentException("해당 수업이 존재하지 않습니다."));
     findClassRoom.update(classRoomDTO);
     classRoomRepository.save(findClassRoom);
   }
 
+  // 교실 정보 조회 (전체)
   public ClassRoomAllInfoDto getAllInfo(UUID classId) {
     ClassRoom classRoom = classRoomRepository.findById(classId)
       .orElseThrow(() -> new IllegalArgumentException("해당 수업이 존재하지 않습니다."));
@@ -118,12 +125,31 @@ public class ClassRoomService {
             .build();
   }
 
+  // 교실 삭제 (선생)
   public void deleteClassRoom(UUID userId, UUID classId) {
+    validateOwner(userId, classId);
     ClassRoom classRoom = findByIdOrElseThrow(classId);
     classRoomRepository.delete(classRoom);
   }
 
+
   public ClassRoom findByIdOrElseThrow(UUID classId){
     return classRoomRepository.findById(classId).orElseThrow(() -> new EntityNotFoundException("해당 교실은 찾을수 없습니다."));
+  }
+
+  // 교실에 대한 선생님 권한 확인
+  public void validateOwner(UUID userId, UUID classId) {
+    ClassRoomUser classRoomUser = (ClassRoomUser) classRoomUserRepository.findByUser_UserIdAndClassRoom_ClassRoomId(userId, classId)
+            .orElseThrow(() -> new AccessDeniedException("해당 교실의 멤버가 아닙니다."));
+
+    if (classRoomUser.getUser().getRole() != Role.TEACHER) {
+      throw new AccessDeniedException("해당 작업을 수행할 권한(교사)이 없습니다.");
+    }
+  }
+
+  // 교실에 속해 있는지 권한 확인
+  public void validateInClassRoom(UUID userId, UUID classId) {
+    ClassRoomUser classRoomUser = (ClassRoomUser) classRoomUserRepository.findByUser_UserIdAndClassRoom_ClassRoomId(userId, classId)
+            .orElseThrow(() -> new AccessDeniedException("해당 교실의 멤버가 아닙니다."));
   }
 }
